% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_estatapi.R
\name{estat_getDataCatalog}
\alias{estat_getDataCatalog}
\alias{estat_getMetaInfo}
\alias{estat_getStatsData}
\alias{estat_getStatsList}
\alias{estat_api}
\title{Wrap functions of \code{{estatapi}} Functions}
\usage{
estat_getDataCatalog(
  lang = c("J", "E"),
  .use_label = TRUE,
  surveyYears = NULL,
  openYears = NULL,
  statsField = NULL,
  statsCode = NULL,
  searchWord = NULL,
  dataType = NULL,
  catalogId = NULL,
  resourceId = NULL,
  startPosition = NULL,
  limit = NULL,
  updatedDate = NULL,
  appId = check_appId(),
  ...
)

estat_getMetaInfo(statsDataId, lang = c("J", "E"), appId = check_appId(), ...)

estat_getStatsData(
  statsDataId,
  startPosition = NULL,
  limit = NULL,
  lang = c("J", "E"),
  .fetch_all = TRUE,
  appId = check_appId(),
  ...
)

estat_getStatsList(
  searchWord,
  lang = c("J", "E"),
  .use_label = TRUE,
  surveyYears = NULL,
  openYears = NULL,
  statsField = NULL,
  statsCode = NULL,
  searchKind = NULL,
  collectArea = NULL,
  startPosition = NULL,
  limit = NULL,
  updatedDate = NULL,
  appId = check_appId(),
  ...
)

estat_api(path, appId = check_appId(), ...)
}
\arguments{
\item{lang}{Language of the data. \code{"J"}(Japanese) or \code{"E"}(English).}

\item{.use_label}{Whether to take the human-redable label value or the code value when
flattening a field containing both (default: \code{TRUE}).}

\item{surveyYears}{Year and month when the survey was conducted. The format is either
\code{"YYYY"}, \code{"YYYYMM"}, or \code{"YYYYMM-YYYYMM"}.}

\item{openYears}{Year and month when the survey result was opened. The format is the
same as \code{surveyYears}.}

\item{statsField}{Field of statistics. The format is either two digits(large classification)
or four digits (small classification). For the detail of the
classification, see \url{http://www.soumu.go.jp/toukei_toukatsu/index/seido/sangyo/26index.htm}.}

\item{statsCode}{Code assigned for each statistical agency and statistics. The format can
be five digits (agency), and eight digits (statistics). For the detail,
see \url{http://www.stat.go.jp/info/guide/public/code/code.htm}.}

\item{searchWord}{Keyword for searching. You can use \code{OR} and \code{AND} to combine conditions
(e.g.: \code{"apple AND orange"}).}

\item{dataType}{Type of data. \code{XLS}: Excel file, \code{CSV}: CSV file, \code{PDF}: PDF file, \code{DB}: Database.}

\item{catalogId}{Catalog ID.}

\item{resourceId}{Catalog resource ID.}

\item{startPosition}{Index of the first record to get.}

\item{limit}{Max number of records to get.}

\item{updatedDate}{Last updated date. The format is either \code{"YYYY"}, \code{"YYYYMM"}, or
\code{"YYYYMM-YYYYMM"}.}

\item{appId}{Application ID.}

\item{...}{Other parameters.
\itemize{
\item \code{lvTab}:
Level of the meta-information. The format can be \code{X} or \code{X-Y}, \code{-X} and \verb{X-}.
\item \code{cdTab}:
Code(s) of the meta-infomation items to select. The format can be a character vector (\code{c("001", "002")}) or
a chraracter of values and commas (\code{"001,002"}).
\item \code{cdTabFrom}:
The code of the first meta-information item to select.
\item \code{cdTabTo}:
The code of the last meta-information item to select.
\item \code{lvTime}:
Level of the time to select. The format is the same as \code{lvTab}
\item \code{cdTime}
Time(s) to select. The format is the same way like \code{cdTab}
\item \code{cdTimeFrom}:
The first time to select. The format is the same way like \code{cdTabFrom}
\item \code{cdTimeTo}:
The last time to select. The format is the same way like \code{cdTabTo}
\item \code{lvArea}:
Level of the area to select. The format is the same as \code{lvTab}
\item \code{cdArea}
Code(s) of the Area to select. The format is the same way like \code{cdTab}
\item \code{cdAreaFrom}:
The code of the first area to select. The format is the same way like \code{cdTabFrom}
\item \code{cdAreaTo}:
The code of the last area to select. The format is the same way like \code{cdTabTo}
\item \code{lvCat01}, \code{cdCat01}, \code{cdCat01From}, \code{cdCat01To}, ...:
The same way like above.
}}

\item{statsDataId}{ID of the statistical dataset.}

\item{.fetch_all}{Whether to fetch all records when the number of records is larger than
100,000.}

\item{searchKind}{Type of statistics.
\itemize{
\item \code{1}:
summary
\item \code{2}:
regional mesh
}}

\item{collectArea}{Area of statistics.
\itemize{
\item \code{1}:
country
\item \code{2}:
prefecture
\item \code{3}:
municipality
}}

\item{path}{API endpoint.}
}
\description{
Wrap function to execute functions of \code{{estatapi}} using the appId set by \code{\link[=set_appId]{set_appId()}}.
If appId is set by \code{\link[=set_appId]{set_appId()}}, the set value is used if not specified by appId in the function.
Also, if appId is specified in the function, it is used first.

Please refer to the \code{{estatapi}} documentation of each function.
}
\seealso{
\code{\link[estatapi:estat_api]{estatapi::estat_api()}}
\code{\link[estatapi:estat_getDataCatalog]{estatapi::estat_getDataCatalog()}}
\code{\link[estatapi:estat_getMetaInfo]{estatapi::estat_getMetaInfo()}}
\code{\link[estatapi:estat_getStatsData]{estatapi::estat_getStatsData()}}
\code{\link[estatapi:estat_getStatsList]{estatapi::estat_getStatsList()}}

\url{http://www.e-stat.go.jp/api/e-stat-manual3-0/#api_2_3}
\url{http://www.e-stat.go.jp/api/e-stat-manual3-0/#api_3_4}
}
