% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appID.R
\name{set_appId}
\alias{set_appId}
\alias{clear_appId}
\alias{check_appId}
\title{e-Stat API ID Tools}
\usage{
set_appId(appId)

clear_appId()

check_appId()
}
\arguments{
\item{appId}{Application ID of e-stat API.}
}
\description{
Functions for setting, deleting, and confirming the API ID (appId) of e-Stat.
}
