% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_appId}
\alias{dir_appId}
\alias{file_appId}
\title{e-Stat API ID Utils}
\usage{
dir_appId()

file_appId()
}
\description{
Function to check the directory and file location where e-Stat's API ID (appId) is stored.
The behavior is independent of the existence of the directory and file.
}
